Scaling - remedy for overflow
We can do it differently but the problem exists and should be taken care
of.

Force the computations to be made on 8 bytes. Otherwise for S32 and U32
we are likely to hit overflow.
* and / are left associative so originally:

*value = ((long long int)*value - mode_info->raw_min)
* (mode_info->si_max - mode_info->si_min)
/ (mode_info->raw_max - mode_info->raw_min)
+ mode_info->si_min;

*value - mode_info->raw_min // ok
* (mode_info->si_max - mode_info->si_min) // this is where we get
overflow

The only reason that the original code works with raw_min,
si_min=INT_MIN, raw_max, si_max=INT_MAX is because:

INT_MAX-INT_MIN is -1 // U2

so we were not getting overflows for high value and this kind of
min/maxes